made stop button work
Reminder for future Seth - every channel send "stacks up", so the receiver processes them 1 at a time, 1 each recv call.

Putting the send in the GUI update loop was generating ~60 messages a second, while the 1 sec sleep in the other thread meant it takes minutes for it to receive the stop signal.